refactor: keep memory test scaffolding and summarizer presets out of the wrong layers
Move SwiftData v1/migration container factories into the test target, colocate MemorySummarizer with InferenceProviderSummarizer, and share one SwiftData sandbox gate.
