+++ Sources/Swarm/Memory/Summarizer.swift
@@ -252,62 +252,3 @@ struct FallbackSummarizer: Summarizer, Sendable {
     private let primary: any Summarizer
     private let fallback: any Summarizer
 }
-
-// MARK: - MemorySummarizer
-
-/// Built-in summarizer presets for ``Memory/summary(configuration:summarizer:)``
-/// and ``Memory/hybrid(configuration:summarizer:)``.
-///
-/// ```swift
-/// // Default factory still truncates (does not summarize).
-/// let truncated: SummaryMemory = .summary()
-///
-/// // Real LLM summarization.
-/// let llm: SummaryMemory = .summary(summarizer: .inferenceProvider(myProvider))
-///
-/// // On-device Foundation Models when available; otherwise truncates.
-/// let onDevice: SummaryMemory = .summary(summarizer: .foundationModels)
-/// ```
-///
-/// These presets exist because the memory factories are synchronous:
-/// they cannot await ``Swarm/defaultProvider`` or Foundation Models
-/// availability the way ``Agent`` does at run time.
-public struct MemorySummarizer: Sendable {
-    /// Truncates (drops) old content. Does **not** summarize.
-    public static var truncating: MemorySummarizer {
-        MemorySummarizer(summarizer: TruncatingSummarizer.shared)
-    }
-
-    /// On-device Foundation Models via
-    /// ``InferenceProviderSummarizer/conversationSummarizer(provider:)``.
-    ///
-    /// Resolves the provider with the same *synchronous* check Agent uses
-    /// for its last fallback (`DefaultInferenceProviderFactory`). If
-    /// Foundation Models are unavailable, this is ``truncating``.
-    /// Environment / ``Swarm/defaultProvider`` are not consulted — those
-    /// require async access that the factory call site does not have.
-    public static var foundationModels: MemorySummarizer {
-        if let provider = DefaultInferenceProviderFactory.makeFoundationModelsProviderIfAvailable() {
-            return MemorySummarizer(
-                summarizer: InferenceProviderSummarizer.conversationSummarizer(provider: provider)
-            )
-        }
-        return .truncating
-    }
-
-    /// LLM summarization using ``InferenceProviderSummarizer/conversationSummarizer(provider:)``.
-    ///
-    /// - Parameter provider: Inference backend used for summarization.
-    /// - Returns: A preset wrapping a conversation summarizer.
-    public static func inferenceProvider(_ provider: any InferenceProvider) -> MemorySummarizer {
-        MemorySummarizer(
-            summarizer: InferenceProviderSummarizer.conversationSummarizer(provider: provider)
-        )
-    }
-
-    let summarizer: any Summarizer
-
-    private init(summarizer: any Summarizer) {
-        self.summarizer = summarizer
-    }
-}
